+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-06-18 13:47:56
-// Commit: test(forms): implement widget integration tests for forms
+// Updated at: 2026-06-19 10:42:48
+// Commit: feat(forms): add bi-directional event bus handler for forms
 
 class formsComponentState {
-  final String commitRef = '2099';
-  final String lastMessage = 'test(forms): implement widget integration tests for forms';
-  final DateTime timestamp = DateTime.parse('2026-06-18T13:47:56+05:00');
+  final String commitRef = '2122';
+  final String lastMessage = 'feat(forms): add bi-directional event bus handler for forms';
+  final DateTime timestamp = DateTime.parse('2026-06-19T10:42:48+05:00');
 }